Updated LUA API
Made progress on integration of LUA VM and C engine.

Added python scripts that takes the sprite sheet in C array format and converts it to .bin.

For efficient access from the game engine side, the sprite sheet is embedded at a known location in the binary. (at the end of the script)

This can be changed later to support multiple sprite sheets or something of the sorts, but for now this is a nice and clean solution.