draw_map functionality
Added draw_map functionality to the API and emulator.

Also changed to concept of "sprite_sheet binary" to game binary, as the single binary that is exported from the vscode extention, now contains the sprite_sheet, the map and the sprite flags all in one.

This still needs to be implemented on the real game engine, but works in emulator for now.